+++ CHANGELOG.md
@@ -22,13 +22,13 @@
 
 - Currency select screen from Menu
 - Change to any base currency to display your BCH balance as.
-  - At first enabling a whistelist of 24 top used currencies
+  - At first enabling a whitelist of 24 top used currencies
 - Number formatting improvements throughout to better match selected currency
 - Improved fee calculations when sending SLP
 
 ### Bug fixes
 
-- Updating bitbox to reduce server load
+- Updating bitbox-sdk to reduce server load
 - Fixed SLP issue sending tokens when user has a large number of SLP related UTXO
 - Fixed bug where new users see negative balance in sending flow
 - Allow sending non fungible tokens without ticker symbols